backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for training the CartPole-v1 balancing bot policy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "array-api-compat>=1.12.0",
    "gymnasium>=1.2.1",
    "hydra-core>=1.3.2",
    "logging>=0.4.9.6",
    "moviepy>=2.2.1",
    "numpy>=2.3.3",
    "onnx>=1.19.0",
    "onnxscript>=0.5.3",
    "pygame>=2.6.1",
    "ruff>=0.13.2",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.1",
]
